Split Items and Removed Weapons
Items are split between Player Items and Blaster Attachments. Blaster moves from Weapons to Items.
