+++ src/PolyGone/Items/BlasterAttachments/MultiShot.cs
@@ -1,6 +1,5 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using PolyGone.Weapons;
 
 namespace PolyGone.Items
 {
