+++ src/PolyGone/Items/BlasterAttachments/PiercingAttachment.cs
@@ -1,6 +1,5 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using PolyGone.Weapons;
 
 namespace PolyGone.Items
 {
