+++ src/PolyGone/Scenes/LoadoutScene.cs
@@ -1,0 +1,6 @@
+namespace PolyGone;
+
+internal class Loadout : IScene
+{
+    
+}
